moved System and SimFunctions to separate submodule
